+++ HISTORY.rst
@@ -21,6 +21,8 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* BSSFL-444: Add notes on serial number
+
 **Bugfixes**
 
 **Build**